Refactor utils to move to folder
