+++ revm/app_item/views.py
@@ -1,7 +1,7 @@
 from rest_framework.permissions import AllowAny
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
-from revm_site.views import CreateResourceViewSet
+from revm_site.utils.views import CreateResourceViewSet
 from .models import Category
 from .serializers import (
     ItemRequestSerializer,
