+++ revm/app_transport_service/admin.py
@@ -4,8 +4,7 @@ from import_export.admin import ImportExportModelAdmin
 
 from app_account.models import CustomUser
 from app_transport_service import models
-from revm_site.admin import CommonRequestInline, CommonOfferInline, CommonResourceAdmin
-from revm_site.utils import CountyFilter
+from revm_site.utils.admin import CommonRequestInline, CommonOfferInline, CommonResourceAdmin, CountyFilter
 
 
 class TransportOfferInline(CommonOfferInline):
