+++ revm/app_transport_service/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from revm_site.serializers import CountyCoverageSerializer
+from revm_site.utils.serializers import CountyCoverageSerializer
 from .models import TransportServiceRequest, TransportServiceOffer, Category
