+++ revm/app_volunteering/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 from app_account.models import CustomUser
-from revm_site.models import (
+from revm_site.utils.models import (
     CommonRequestModel,
     CommonMultipleCountyModel,
     CommonOfferModel,
@@ -10,7 +10,7 @@ from revm_site.models import (
     CommonTransportableModel,
     CommonLocationModel,
 )
-from revm_site.validators import validate_date_disallow_past
+from revm_site.utils.validators import validate_date_disallow_past
 
 
 class Type(models.Model):
