+++ revm/app_volunteering/views.py
@@ -1,10 +1,8 @@
 from rest_framework.permissions import AllowAny
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
-from revm_site.views import CreateResourceViewSet
-
+from revm_site.utils.views import CreateResourceViewSet
 from .models import Type
-
 from .serializers import VolunteeringRequestSerializer, VolunteeringOfferSerializer, VolunteeringCategorySerializer
