+++ revm/revm_site/utils/admin.py
@@ -1,3 +1,4 @@
+from django.conf import settings
 from django.contrib import admin, messages
 from django.utils.translation import gettext_lazy as _
 from import_export.admin import ImportExportModelAdmin
@@ -69,3 +70,26 @@ class CommonResourceSingleCountyAdmin(ImportExportModelAdmin):
 
     class Meta:
         abstract = True
+
+
+class CountyFilter(admin.SimpleListFilter):
+    title = _("County")
+
+    parameter_name = "county"
+
+    def lookups(self, request, model_admin):
+        return settings.COUNTY_CHOICES
+
+    def queryset(self, request, queryset):
+        """
+        Returns the filtered queryset based on the value
+        provided in the query string and retrievable via
+        `self.value()`.
+        """
+        # Compare the requested value (either '80s' or '90s')
+        # to decide how to filter the queryset.
+        if self.value():
+            return queryset.filter(
+                county_coverage__contains=self.value(),
+            )
+        return queryset
